clang-tidy: small enum switches to if; drop redundant casts
Convert 1-3 branch enum switches (which trip -Wswitch when cases are
incomplete) to if/else chains. Remove readability-redundant-casting
hits where the expression already has the target type.

prompt: check warning for clang -Wswitch and convert those that only check 1-3 branches to if; fix (clang-tidy readability-redundant-casting)

+++ src/base/File_win.cpp
@@ -1223,7 +1223,7 @@ bool CreateAll(Str dir, int* errOut) {
     if (Exists(dir)) {
         return true;
     }
-    int err = (int)SHCreateDirectoryExW(nullptr, CWStrTemp(dir), nullptr);
+    int err = SHCreateDirectoryExW(nullptr, CWStrTemp(dir), nullptr);
     if (errOut) {
         *errOut = err;
     }
