+++ src/wingui/DialogSizer.cpp
@@ -24,7 +24,7 @@ class DialogData {
 
         // Copy all of the user controls etc. for later, this way the user can quite happily
         // let the structure go out of scope.
-        this->psd = (DialogSizerSizingItem*)memdup((void*)psd, (int)nItemCount * sizeofi(DialogSizerSizingItem));
+        this->psd = (DialogSizerSizingItem*)memdup((void*)psd, nItemCount * sizeofi(DialogSizerSizingItem));
         if (!this->psd) {
             nItemCount = 0;
         }
